refactor trap relative functions
1. Add initialization part into boot.S
2. The change monitor.S to vector.S
3. Delete the term of trap
  "Please do not get confused with trap and exception."
