--- hypervisor/monitor.h
@@ -1,8 +1,0 @@
-#ifndef __MONITOR_H__
-#define __MONITOR_H__
-
-void mon_init(void);
-void mon_smc(void);
-void mon_enter_hyp(void);
-
-#endif
