+++ hypervisor/vector.S
@@ -11,23 +11,23 @@
     .text
 
 /* ---[Secure Mode]------------------------------------------------------ */
-
 /*
  * Secure Monitor Vector Table
  */
+ .global monitor_secure_vectors
 .align 5
     monitor_secure_vectors:
 .word 0 /* reset */
-    b   trap_secure_unhandled   /* undef*/
-    b   trap_smc                /* smc */
-    b   trap_secure_unhandled   /* pabt*/
-    b   trap_secure_unhandled   /* dabt */
-    b   trap_hyp_entry          /* hvc */
-    b   trap_secure_unhandled   /* irq */
-    b   trap_secure_unhandled   /* fiq*/
+    b   vector_secure_unhandled   /* undef*/
+    b   vector_smc                /* smc */
+    b   vector_secure_unhandled   /* pabt*/
+    b   vector_secure_unhandled   /* dabt */
+    b   vector_hyp_entry          /* hvc */
+    b   vector_secure_unhandled   /* irq */
+    b   vector_secure_unhandled   /* fiq*/
 
 /* Install hvbar, hyp vector, and return in Non-secure state */
-trap_smc:
+vector_smc:
     mrc p15, 0, r10, c1, c1, 0  @ SCR -> r10
     @ SCR.NS=1, IRQ,FIQ,EA=0, FW,AW=1, nET=0, SCD=1, HCE=1
     bic r10, r10, #0x07f
@@ -42,14 +42,14 @@ trap_smc:
     @ return in NS state
     movs    pc, lr
 
-trap_hyp_entry:
+vector_hyp_entry:
     /* Setup stack for Hyp for the first time */
     ldr sp, = mon_stacklimit
     /* Stay in Hyp mode and branch to where hvc invoked. */
     mrs lr, elr_hyp
     mov pc, lr
 
-trap_secure_unhandled:
+vector_secure_unhandled:
 @ Push registers
     push    {r0 - r12}
     mrs r0, spsr_mon
@@ -57,42 +57,14 @@ trap_secure_unhandled:
     push    {r0, r1}
     mov r0, sp
     /* Place holder in case we need to handle secure state exception */
-    /* bl   _trap_secure_unhandled */
+    /* bl   _vector_secure_unhandled */
     @ Pop registers
     pop     {r0 - r1}
     msr spsr_mon, r0
     pop     {r0 - r12}
     eret
-/*
- * Initialize vector and stack pointers for modes:
- * Secure SVC, Non-secure SVC, and Hyp mode
- */
-.global mon_init
-    mon_init:
-    push {r0 - r2, lr}
-    @ Setup Exception Vector
-    ldr r1, = monitor_secure_vectors
-    mcr p15, 0, r1, c12, c0, 1      @ Monitor vector base address
-    @ regular return
-    pop {r0 - r2, pc}
 
-.type mon_init, % function
-
-.global mon_enter_hyp
-    mon_enter_hyp:
-    smc #0  @ -> trap_smc
-    hvc #0  @ -> trap_hyp_entry
-    /* Place the right Hyp exception vector for the first time */
-    ldr r0, = monitor_hyp_vectors
-    mcr p15, 4, r0, c12, c0, 0
-    bl main
-    @ never come back
-    1:
-    b 1b
-
-.type mon_enter_hyp, % function
-
-
+/* ---[Hyp Mode]------------------------------------------------------ */
 .global monitor_hyp_vectors
 /*
  * Monitor Vector Table
@@ -100,16 +72,16 @@ trap_secure_unhandled:
 .align 5
 monitor_hyp_vectors:
     .word 0    /* reset */
-    b    trap_unhandled    /* undef*/
-    b    trap_unhandled     /* hypervisor */
-    b    trap_unhandled    /* pabt*/
-    b    trap_dabt    /* dabt */
-    b    trap_hvc     /* hvc */
-    b    trap_irq    /* irq */
-    b    trap_unhandled    /* fiq*/
+    b    vector_unhandled    /* undef*/
+    b    vector_unhandled     /* hypervisor */
+    b    vector_unhandled    /* pabt*/
+    b    vector_dabt    /* dabt */
+    b    vector_hvc     /* hvc */
+    b    vector_irq    /* irq */
+    b    vector_unhandled    /* fiq*/
 
 
-trap_hvc:
+vector_hvc:
     @ Push registers
     push    {r0-r12}
     mrs    r0, spsr_hyp
@@ -118,7 +90,7 @@ trap_hvc:
 
     @ service other argument values -> _hyp_hvc_service(sp)
     mov    r0, sp
-    bl    _hyp_hvc_service    @ r0: HSR
+    bl    _hyp_hvc    @ r0: HSR
 
     @ r0 = return
     tst    r0, #1
@@ -145,16 +117,16 @@ trap_hvc:
     mrs    lr, elr_hyp
     mov    pc, lr
 
-trap_dabt:
+vector_dabt:
     @ Push registers
     push    {r0-r12}
     mrs    r0, spsr_hyp
     mrs    r1, elr_hyp
     push    {r0, r1, lr}
 
-    @ service other argument values -> _hyp_trap_dabort(sp)
+    @ service other argument values -> _hyp_vector_dabort(sp)
     mov    r0, sp
-    bl    _hyp_trap_dabort    @ r0: HSR
+    bl    _hyp_dabort    @ r0: HSR
 
     @ if (return == 0) execute a conditional code
     @ tst    r0, #1
@@ -167,16 +139,16 @@ trap_dabt:
     pop     {r0-r12}
     eret
 
-trap_irq:
+vector_irq:
     @ Push registers
     push    {r0-r12}
     mrs    r0, spsr_hyp
     mrs    r1, elr_hyp
     push    {r0, r1, lr}
 
-    @ service other argument values -> _hyp_trap_irq(sp)
+    @ service other argument values -> _hyp_vector_irq(sp)
     mov    r0, sp
-    bl    _hyp_trap_irq    @ r0: HSR
+    bl    _hyp_irq    @ r0: HSR
 
     @ Pop registers
     pop     {r0-r1, lr}
@@ -185,16 +157,16 @@ trap_irq:
     pop     {r0-r12}
     eret
 
-trap_unhandled:
+vector_unhandled:
     @ Push registers
     push    {r0-r12}
     mrs     r0, spsr_hyp
     mrs     r1, elr_hyp
     push    {r0, r1, lr}
 
-    @ service other argument values -> _hyp_trap_unhandled(sp)
+    @ service other argument values -> _hyp_vector_unhandled(sp)
     mov    r0, sp
-    bl    _hyp_trap_unhandled    @ r0: HSR
+    bl    _hyp_unhandled    @ r0: HSR
 
     @ Pop registers
     pop     {r0-r1, lr}
@@ -202,16 +174,3 @@ trap_unhandled:
     msr     elr_hyp, r1
     pop     {r0-r12}
     eret
-
-.global __mon_switch_to_guest_context
-__mon_switch_to_guest_context:    @ struct arch_regs *r0
-    ldmfd    r0!, {r1, r3, lr}
-    msr     spsr_hyp, r1
-    msr     elr_hyp, r3
-    ldm     r0, {r0-r12}
-    clrex
-    eret
-.type __mon_switch_to_guest_context, %function
-
-
-
