+++ platform-device/cortex_a15x2_rtsm/boot.S
@@ -192,19 +192,39 @@ start:
 	@ Never reached
 1:	b 1b
 
-	@
-	@ Function for C code to make semihosting calls:
-	@
-	.globl __semi_call
-__semi_call:
-#if defined(MACH_MPS)
-	@ M profile semihosting is via bpkt
-	bkpt    0xab
-#elif defined(__thumb__)
-	@ Otherwise, different SVC numbers for ARM or Thumb mode
-	svc    0xab
-#else
-	svc     0x123456
-#endif
-	mov pc, lr
+/*
+ * Initialize vector and stack pointers for modes:
+ * Secure SVC, Non-secure SVC, and Hyp mode
+ */
+.global mon_init
+    mon_init:
+    push {r0 - r2, lr}
+    @ Setup Exception Vector
+    ldr r1, = monitor_secure_vectors
+    mcr p15, 0, r1, c12, c0, 1      @ Monitor vector base address
+    @ regular return
+    pop {r0 - r2, pc}
+.type mon_init, % function
 
+.global mon_enter_hyp
+    mon_enter_hyp:
+    smc #0  @ -> trap_smc
+    hvc #0  @ -> trap_hyp_entry
+    /* Place the right Hyp exception vector for the first time */
+    ldr r0, = monitor_hyp_vectors
+    mcr p15, 4, r0, c12, c0, 0
+    bl main
+    @ never come back
+    1:
+    b 1b
+.type mon_enter_hyp, % function
+
+.global __mon_switch_to_guest_context
+__mon_switch_to_guest_context:    @ struct arch_regs *r0
+    ldmfd    r0!, {r1, r3, lr}
+    msr     spsr_hyp, r1
+    msr     elr_hyp, r3
+    ldm     r0, {r0-r12}
+    clrex
+    eret
+.type __mon_switch_to_guest_context, %function
